Revert "Features/bookmark"
