+++ public/robots.txt
@@ -1,0 +1,3 @@
+User-agent: *
+Allow: /
+Sitemap: https://content-creation-tool.vercel.app/sitemap.xml
