+++ src/app/(app)/tools/[id]/product-detail.tsx
@@ -14,7 +14,7 @@ type Props = {
 
 export default function ProductDetail({ params }: Props) {
   const id = useSearchParams().get("id");
-  const { productList } = useSelector((state: RootState) => state.products);
+  const { productList } = useSelector((state: RootState) => state.product);
   const [product, setProductData] = useState<AirtableModel>();
   const { setVisibleItem } = useVisibleItemContextData();
   const currentCategory = product && product!.fields.Tags[0];
@@ -28,6 +28,7 @@ export default function ProductDetail({ params }: Props) {
       return formattedTitle === slug;
     });
 
+    console.log("productMatched", productMatched);
     if (productMatched) {
       setProductData(productMatched);
     }
@@ -43,7 +44,6 @@ export default function ProductDetail({ params }: Props) {
 
   return (
     <>
-      {" "}
       {product && (
         <>
           <ProductToolBanner
