+++ src/app/api/tools/route.ts
@@ -1,4 +1,4 @@
-export const dynamic = "force-dynamic";
+// export const dynamic = "force-dynamic";
 import { NextResponse } from "next/server";
 import { AirtableModel } from "@/models/airtable.model";
 import axios from "axios";
