+++ src/components/filter/FilterSection.tsx
@@ -32,19 +32,24 @@ export default function FilterSection() {
   /*Redux Dispatch & Selector*/
   const dispatch = useDispatch();
   const categoryData = useSelector(
-    (store: RootState) => store.categories.categoryData
+    (store: RootState) => store.category.categoryData
   );
   const searchQuery = useSelector(
     (store: RootState) => store.search.searchQuery
   );
-  const { productList } = useSelector((state: RootState) => state.products);
+  const filterData = useSelector(
+    (store: RootState) => store.search.searchFilterList
+  );
+  const { productList } = useSelector((state: RootState) => state.product);
   const searchToFocusInput = useSelector(
     (state: RootState) => state.search.searchToFocus
   );
   const scrollPosition = useSelector(
     (state: RootState) => state.search.scrollPosition
   );
-  const { priceData } = useSelector((state: RootState) => state.pricingModel);
+  const { priceData, matchedPrice } = useSelector(
+    (state: RootState) => state.priceModel
+  );
 
   /*Context Data*/
   const { setVisibleItem } = useVisibleItemContextData();
@@ -53,7 +58,7 @@ export default function FilterSection() {
   const handleSearch = (event: React.ChangeEvent<HTMLInputElement>) => {
     // console.log("searchRef", searchRef.current?.value);
     router.push("/");
-    dispatch(scrollPage(600));
+    dispatch(scrollPage(600))
     dispatch(clearMatchedPrice());
     dispatch(clearPriceData());
     dispatch(clearCategoryData());
@@ -156,7 +161,7 @@ export default function FilterSection() {
   };
 
   const categoryOptionsList = Array.from(getListOfCategory()).map(
-    (item: string) => {
+    (item: string, index: number) => {
       return {
         value: item,
         label: item,
@@ -175,12 +180,23 @@ export default function FilterSection() {
     return priceItem;
   };
 
-  const priceOptionList = Array.from(priceModelList()).map((item: string) => {
-    return {
-      value: item,
-      label: item,
-    };
-  });
+  const priceOptionList = Array.from(priceModelList()).map(
+    (item: string, index: number) => {
+      return {
+        value: item,
+        label: item,
+      };
+    }
+  );
+
+  const priceTypeHandler = useCallback(() => {
+    const getPriceList = productList.filter(
+      (item: AirtableModel) =>
+        item?.fields?.Pricing[0]?.toLowerCase() == priceData?.toLowerCase()
+    );
+    dispatch(setMatchedPrice(getPriceList));
+    console.log(getPriceList);
+  }, []);
 
   useEffect(() => {
     SetIsMounted(true);
