+++ src/components/prompt/PromptCard.tsx
@@ -11,13 +11,25 @@ import { Button } from "../ui/button";
 import { useRouter } from "next/navigation";
 import { FiArrowUpRight } from "react-icons/fi";
 import { Badge } from "../ui/badge";
-import LikeButton from "../ui/likebutton";
-import BookmarkButton from "../ui/bookmarkbutton";
+import { AiFillHeart, AiOutlineHeart } from "react-icons/ai";
+import { BsBookmark, BsBookmarkFill } from "react-icons/bs";
 
 const PromptCard = ({
-  promptResource,
+  promptResource = {
+    id: "recEyJBryU2v9u5KM",
+    fields: {
+      Source: "GPTBot",
+      Name: "Create a social media schedule",
+      Category: ["Content Creation"],
+      Description:
+        "As a skilled social media manager, your task is to formulate a one-month social media posting calendar beginning from [insert date that schedule will start]. The frequency of posts should align with your preference, either [daily/every two days/every weekday/weekly]. Your business, known as [insert name], is engaged in selling [insert products or services]. Each post should detail the publish date, an engaging heading, captivating body text, and relevant hashtags. Ensure that the tone is [professional/casual/funny/friendly] as per our brand’s voice. Additionally, each post should include a suggestion for a suitable image, which could be sourced from a stock image service.",
+      Tags: ["LinkedIn", "Instagram", "Twitter"],
+      Status: "",
+      SourceLink: "",
+    },
+  },
 }: {
-  promptResource: PropmtResourceModel;
+  promptResource?: PropmtResourceModel;
 }) => {
   const router = useRouter();
   const [isBookMarked, setIsBookMarked] = useState<boolean>(false);
@@ -27,12 +39,17 @@ const PromptCard = ({
     const title = promptResource?.fields?.Name?.toLowerCase()
       .trim()
       .replace(/\s/g, "-");
-    router.push(`/prompt/${title}`);
-  };
 
-  if (!promptResource) {
-    return null;
-  }
+    router.push(`/prompt/${title}`);
+
+    console.log("PromptCard handlePromptClick", title);
+  };
+  const handleLike = async () => {
+    setIsLiked(!isLiked);
+  };
+  const handleBoomark = async () => {
+    setIsBookMarked(!isBookMarked);
+  };
   return (
     <Card className="flex flex-col border-black hover:border-DarkOrange bg-light-gray ">
       <CardHeader>
@@ -44,7 +61,7 @@ const PromptCard = ({
               </Badge>
             ))}
           </div>
-          {/* <button
+          <button
             title="Bookmark"
             type="button"
             className="flex items-center gap-x-2"
@@ -55,15 +72,8 @@ const PromptCard = ({
             ) : (
               <AiOutlineHeart className="text-2xl   text-black" />
             )}
-            <span>{Math.floor(Math.random() * 10) + 1}</span>
-          </button> */}
-          <LikeButton
-            key={promptResource.id}
-            initialLikedState={isLiked}
-            itemId={promptResource.id}
-            itemName={promptResource.fields?.Name}
-            itemType="prompt"
-          />
+            <span>23</span>
+          </button>
         </div>
         <CardTitle className="text-lg">{promptResource.fields?.Name}</CardTitle>
       </CardHeader>
@@ -81,20 +91,13 @@ const PromptCard = ({
             View Prompt
             <FiArrowUpRight className="text-2xl " />
           </Button>
-          {/* <button title="Bookmark" onClick={handleBoomark}>
+          <button title="Bookmark" onClick={handleBoomark}>
             {isBookMarked ? (
               <BsBookmarkFill className="text-2xl text-DarkOrange" />
             ) : (
               <BsBookmark className="text-2xl  text-black" />
             )}
-          </button> */}
-          <BookmarkButton
-            key={promptResource.id}
-            isInitialBookmarked={isBookMarked}
-            Name={promptResource.fields?.Name}
-            id={promptResource.id}
-            itemType="prompt"
-          />
+          </button>
         </div>
       </CardFooter>
     </Card>
