+++ src/components/prompt/prompt-library.tsx
@@ -14,16 +14,13 @@ import Pagination from "../pagination/Pagination";
 
 export default function PromptLibrary() {
   const [currentPage, setCurrentPage] = useState(1);
+
   const [isCategoryOpen, setIsCategoryOpen] = useState(false);
   const [promptList, setPromptList] = useState<PropmtResourceModel[]>([]);
-  const [filteredPrompts, setFilteredPrompts] = useState<PropmtResourceModel[]>(
-    []
-  );
-  const [selectedCategory, setSelectedCategory] = useState<string | null>(null);
 
   const dispatch: AppDispatch = useDispatch();
   const { isLoading, isError, promptResourceList } = useSelector(
-    (state: RootState) => state.promptResources
+    (state: RootState) => state.promptResource
   );
 
   const itemsPerPage = 10;
@@ -32,36 +29,29 @@ export default function PromptLibrary() {
   };
 
   // eslint-disable-next-line react-hooks/exhaustive-deps
-  const getPromptCategory = useCallback(() => {
+  const getPromptCategory = () => {
     const categoryList = new Set();
     if (promptList.length > 0) {
-      promptList.forEach((prompt) => {
-        if (prompt.fields?.Category?.[0]) {
-          categoryList.add(prompt.fields.Category[0]);
+      promptList.map((prompt) => {
+        if (!categoryList?.has(prompt.fields?.Category?.[0])) {
+          categoryList.add(prompt.fields?.Category?.[0]);
         }
       });
     }
-    return Array.from(categoryList) as string[];
-  }, [promptList]);
 
-  const handleCategorySelect = (category: string) => {
-    setSelectedCategory(category);
-    filterPrompts(category);
-    setCurrentPage(1);
-  };
-  const filterPrompts = (category: string) => {
-    const filtered = promptList.filter(
-      (prompt) => prompt.fields?.Category?.[0] === category
-    );
-    setFilteredPrompts(filtered);
+    return Array.from(categoryList);
   };
 
   const updateCurrentProducts = useCallback(() => {
     const startIndex = (currentPage - 1) * itemsPerPage;
     const endIndex = startIndex + itemsPerPage;
-    const currentList = selectedCategory ? filteredPrompts : promptList;
-    return currentList.slice(startIndex, endIndex);
-  }, [currentPage, promptList, filteredPrompts, selectedCategory]);
+    const newCurrentProducts = promptList!.slice(startIndex, endIndex);
+    return newCurrentProducts;
+  }, [currentPage, promptList]);
+
+  useEffect(() => {
+    updateCurrentProducts();
+  }, [currentPage, promptList, updateCurrentProducts]);
 
   useEffect(() => {
     if (promptResourceList?.length === 0) {
@@ -70,11 +60,11 @@ export default function PromptLibrary() {
   }, [dispatch, promptResourceList]);
 
   useEffect(() => {
+    console.log("useEffect Triggered!!!");
     if (!isLoading && promptResourceList?.length > 0) {
       setPromptList(promptResourceList);
-      setFilteredPrompts(promptResourceList);
     }
-  }, [isLoading, promptResourceList, isError]);
+  }, [isLoading, promptResourceList, isError, getPromptCategory]);
 
   if (isLoading) {
     return <Loader />;
@@ -105,11 +95,7 @@ export default function PromptLibrary() {
               isCategoryOpen ? "block" : "hidden"
             } lg:block`}
           >
-            <PromptCategory
-              categoryList={getPromptCategory()}
-              onSelectCategory={handleCategorySelect}
-              selectedCategory={selectedCategory}
-            />
+            <PromptCategory categoryList={getPromptCategory()} />
           </aside>
 
           {/* Main Content */}
@@ -127,9 +113,7 @@ export default function PromptLibrary() {
               </div>
               <Pagination
                 currentPage={currentPage}
-                totalItems={
-                  selectedCategory ? filteredPrompts.length : promptList.length
-                }
+                totalItems={promptList!.length}
                 onPageChange={handlePageChange}
               />
             </main>
