--- src/hooks/useDebounce.ts
@@ -1,20 +1,0 @@
-import { useRef, useEffect } from "react";
-
-export function useDebounce(callback: Function, delay: number) {
-  const timeoutRef = useRef<NodeJS.Timeout | null>(null);
-
-  useEffect(() => {
-    return () => {
-      if (timeoutRef.current) clearTimeout(timeoutRef.current);
-    };
-  }, []);
-
-  const debounce = (...args: any[]) => {
-    if (timeoutRef.current) clearTimeout(timeoutRef.current);
-    timeoutRef.current = setTimeout(() => {
-      callback(...args);
-    }, delay);
-  };
-
-  return debounce;
-}
