+++ src/models/bookmark/Bookmark.model.ts
@@ -2,24 +2,17 @@ import mongoose, { Schema } from "mongoose";
 
 export interface Bookmark extends Document {
   userId: Schema.Types.ObjectId;
-  itemId: string;
-  itemType: "tool" | "prompt";
+  productId: string;
 }
 
 const bookmarkSchema: Schema<Bookmark> = new Schema({
   userId: {
     type: Schema.Types.ObjectId,
     ref: "User",
-    required: [true, "UserId is required"],
   },
-  itemId: {
+  productId: {
     type: String,
-    required: [true, "ItemId is required"],
-  },
-  itemType: {
-    type: String,
-    enum: ["tool", "prompt"],
-    required: [true, "Item type is required"],
+    required: [true, "Product id is required"],
   },
 });
