--- src/redux/slice/bookmark/boookmark-old..slicce..tts
@@ -1,143 +1,0 @@
-
-import { createSlice, createAsyncThunk, PayloadAction } from "@reduxjs/toolkit";
-
-// Define types for bookmark items and state
-export interface BookmarkItem {
-  itemId: string;
-  itemType: string;
-}
-
-interface BookmarkState {
-  isBookmarkChecked: boolean;
-  bookmarkList: BookmarkItem[];
-  status: "idle" | "loading" | "succeeded" | "failed";
-  getListStatus: "idle" | "loading" | "succeeded" | "failed";
-  error: string | null;
-}
-
-// Async thunk to fetch the bookmark list
-export const getBookmarkList = createAsyncThunk<BookmarkItem[]>(
-  "bookmark/fetchBookmarkList",
-  async () => {
-    const response = await fetch("/api/bookmark");
-    const jsonData = await response.json();
-    return jsonData?.bookmarks[0]?.items; // Update this according to your response structure
-  }
-);
-
-// Async thunk to add a bookmark
-export const addBookmark = createAsyncThunk<
-  BookmarkItem,
-  { itemId: string; itemType: string }
->("bookmark/addBookmark", async ({ itemId, itemType }, { dispatch }) => {
-  const response = await fetch("/api/bookmark", {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({ itemId, itemType }),
-  });
-
-  dispatch(getBookmarkList());
-  if (!response.ok) {
-    throw new Error("Failed to add bookmark");
-  }
-  return { itemId, itemType };
-});
-
-// Async thunk to delete a bookmark
-export const deleteBookmark = createAsyncThunk<
-  BookmarkItem,
-  { itemId: string; itemType: string }
->("bookmark/deleteBookmark", async ({ itemId, itemType }, { dispatch }) => {
-  const response = await fetch("/api/bookmark", {
-    method: "DELETE",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({ itemId, itemType }),
-  });
-
-  dispatch(getBookmarkList());
-  if (!response.ok) {
-    throw new Error("Failed to delete bookmark");
-  }
-  return { itemId, itemType };
-});
-
-// Initial state
-const initialState: BookmarkState = {
-  isBookmarkChecked: false,
-  bookmarkList: [],
-  status: "idle",
-  getListStatus: "idle",
-  error: null,
-};
-
-// Create the slice
-const bookmarkSlice = createSlice({
-  name: "bookmark",
-  initialState,
-  reducers: {
-    setIsBookmarkCheck: (state) => {
-      state.isBookmarkChecked = !state.isBookmarkChecked;
-    },
-    clearBookmarkList: (state) => {
-      state.bookmarkList = [];
-    },
-    updateBookmarkList(state, action: PayloadAction<BookmarkItem[]>) {
-      state.bookmarkList = action.payload;
-    },
-  },
-  extraReducers: (builder) => {
-    builder
-      .addCase(getBookmarkList.pending, (state) => {
-        state.getListStatus = "loading";
-      })
-      .addCase(
-        getBookmarkList.fulfilled,
-        (state, action: PayloadAction<BookmarkItem[]>) => {
-          state.getListStatus = "succeeded";
-          state.bookmarkList = action.payload;
-        }
-      )
-      .addCase(getBookmarkList.rejected, (state, action) => {
-        state.getListStatus = "failed";
-        state.error = action.error.message || null;
-      })
-      .addCase(addBookmark.pending, (state) => {
-        state.status = "loading";
-      })
-      .addCase(
-        addBookmark.fulfilled,
-        (state, action: PayloadAction<BookmarkItem>) => {
-          state.status = "succeeded";
-          state.bookmarkList.push(action.payload);
-        }
-      )
-      .addCase(addBookmark.rejected, (state, action) => {
-        state.status = "failed";
-        state.error = action.error.message || null;
-      })
-      .addCase(deleteBookmark.pending, (state) => {
-        state.status = "loading";
-      })
-      .addCase(
-        deleteBookmark.fulfilled,
-        (state, action: PayloadAction<BookmarkItem>) => {
-          state.status = "succeeded";
-          state.bookmarkList = state.bookmarkList.filter(
-            (bookmark) =>
-              bookmark.itemId !== action.payload.itemId ||
-              bookmark.itemType !== action.payload.itemType
-          );
-        }
-      )
-      .addCase(deleteBookmark.rejected, (state, action) => {
-        state.status = "failed";
-        state.error = action.error.message || null;
-      });
-  },
-});
-
-// Export actions and reducer
-export const { setIsBookmarkCheck, clearBookmarkList, updateBookmarkList } =
-  bookmarkSlice.actions;
-export default bookmarkSlice.reducer;
-s
