--- src/redux/slice/resources/resource.slice.ts
@@ -1,55 +1,0 @@
-import { ResourceModel } from "@/models/airtable.model";
-import { createSlice, createAsyncThunk, PayloadAction } from "@reduxjs/toolkit";
-
-// Define the shape of the initial state
-interface ResourceState {
-  isLoading: boolean;
-  error: string | null;
-  resourceList: ResourceModel[];
-}
-
-// Define the initial state
-const initialState: ResourceState = {
-  isLoading: false,
-  error: null,
-  resourceList: [],
-};
-
-// Async thunk to fetch data from the Airtable list
-export const fetchResourcesList = createAsyncThunk<
-  ResourceState["resourceList"],
-  void
->("fetch/resourceList", async () => {
-  const response = await fetch("/api/resources");
-  const responseBody = await response.json();
-  return responseBody.data;
-});
-
-// Create the slice
-const resourceSlice = createSlice({
-  name: "resourceList",
-  initialState,
-  reducers: {},
-  extraReducers: (builder) => {
-    builder.addCase(
-      fetchResourcesList.fulfilled,
-      (state, action: PayloadAction<ResourceState["resourceList"]>) => {
-        state.isLoading = false;
-        state.resourceList = action.payload;
-      }
-    );
-    builder.addCase(fetchResourcesList.pending, (state) => {
-      state.isLoading = true;
-    });
-    builder.addCase(fetchResourcesList.rejected, (state, action) => {
-      state.isLoading = false;
-      state.error = action.error.message ?? "Something went wrong";
-    });
-  },
-});
-
-// Export the actions (if you have defined any in reducers)
-export const {} = resourceSlice.actions;
-
-// Export the reducer
-export default resourceSlice.reducer;
