+++ src/redux/store.ts
@@ -1,28 +1,26 @@
 import { configureStore } from "@reduxjs/toolkit";
-import productSlice from "@/redux/slice/product/product.slice";
-import categorySlice from "@/redux/slice/category/category.slice";
-import verifiedProductSlice from "@/redux/slice/verified/verified.slice";
-import bookmarkSlice from "@/redux/slice/bookmark/bookmark.slice";
-import resourceSlice from "@/redux/slice/resources/resource.slice";
-import upvoteSlice from "@/redux/slice/upvote/upvoteSlice";
-import priceModelSlice from "@/redux/slice/priceModel/priceModel.slice";
-import promptResourceSlice from "@/redux/slice/prompt-resource/promptResource.slice";
-import authSlice from "@/redux/slice/auth/auth.slice";
-import searchSlice from "@/redux/slice/search/search.slice";
-import likeSlice from "./slice/like/like.slice";
+import userSlice from "@/redux/slice/user/userSlice";
+import productSlice from "@/redux/slice/product/productSlice";
+import bookmarkSlice from "@/redux/slice/bookmark/bookmarkSlice";
+import categorySlice from "@/redux/slice/category/categorySlice";
+import searchSlice from "@/redux/slice/search/searchSlice";
+import verifiedProductSlice from "@/redux/slice/verified/verifiedSlice";
+import upvoteSlice from "./slice/upvote/upvoteSlice";
+import priceModelSlice from "./slice/priceModel/priceModelSlice";
+import promptResourceSlice from "./slice/prompt-resource/promptResourceSlice";
+import resourceSlice from "./slice/resources/ResourcesSlice";
 
 const appStore = configureStore({
   reducer: {
-    auth: authSlice,
-    products: productSlice,
-    bookmarks: bookmarkSlice,
-    categories: categorySlice,
+    user: userSlice,
+    product: productSlice,
+    bookmark: bookmarkSlice,
+    category: categorySlice,
     search: searchSlice,
-    verifiedProducts: verifiedProductSlice,
-    upvotes: upvoteSlice,
-    likes: likeSlice,
-    pricingModel: priceModelSlice,
-    promptResources: promptResourceSlice,
+    verifiedProduct: verifiedProductSlice,
+    upvote: upvoteSlice,
+    priceModel: priceModelSlice,
+    promptResource: promptResourceSlice,
     resources: resourceSlice,
   },
 });
